make up and down actions idempotent

+++ resources/application.rb
@@ -29,10 +29,12 @@ end
 
 action :up do
   project_name = new_resource.project_name || current_resource.project_name
+  compose_files = new_resource.compose_files || current_resource.compose_files
 
   execute "running docker-compose up for project #{project_name}" do
     command "docker-compose #{get_compose_params} up #{get_up_params}"
     environment('PATH' => '/usr/local/sbin:/usr/local/bin:/sbin:/bin:/usr/sbin:/usr/bin')
+    only_if "[ $(docker-compose -f #{compose_files.join(' -f ')} ps -q | wc -l) -eq 0 ]"
     user 'root'
     group 'root'
   end
@@ -51,10 +53,12 @@ end
 
 action :down do
   project_name = new_resource.project_name || current_resource.project_name
+  compose_files = new_resource.compose_files || current_resource.compose_files
 
   execute "running docker-compose down for project #{project_name}" do
     command "docker-compose #{get_compose_params} down  #{get_down_params}"
     environment('PATH' => '/usr/local/sbin:/usr/local/bin:/sbin:/bin:/usr/sbin:/usr/bin')
+    not_if "[ $(docker-compose -f #{compose_files.join(' -f ')} ps -q | wc -l) -eq 0 ]"
     user 'root'
     group 'root'
   end
